Allow removing members from the route relation.
