+++ src/main/java/org/openstreetmap/josm/plugins/pt_assistant/gui/routing/RouteTraverser.java
@@ -41,6 +41,10 @@ public class RouteTraverser {
         this.segments = findSegments();
     }
 
+    public RouteType getType() {
+        return type;
+    }
+
     public List<RouteStop> findStopPositions() {
         List<List<RelationMember>> groupedByStop = groupByStop();
