+++ src/main/java/org/openstreetmap/josm/plugins/pt_assistant/gui/routing/RoutingPanel.java
@@ -3,14 +3,12 @@ package org.openstreetmap.josm.plugins.pt_assistant.gui.routing;
 import static org.openstreetmap.josm.tools.I18n.tr;
 
 import java.awt.BorderLayout;
-import java.awt.Component;
 import java.awt.Point;
 import java.awt.event.ActionEvent;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 
-import javax.swing.JButton;
 import javax.swing.JComponent;
 import javax.swing.JMenuItem;
 import javax.swing.JPopupMenu;
@@ -18,22 +16,19 @@ import javax.swing.JSeparator;
 import javax.swing.border.EmptyBorder;
 
 import org.openstreetmap.josm.actions.JosmAction;
-import org.openstreetmap.josm.data.osm.Node;
 import org.openstreetmap.josm.data.osm.OsmPrimitive;
 import org.openstreetmap.josm.data.osm.Relation;
-import org.openstreetmap.josm.data.osm.Way;
 import org.openstreetmap.josm.gui.dialogs.relation.actions.IRelationEditorActionAccess;
-import org.openstreetmap.josm.plugins.customizepublictransportstop.OSMTags;
-import org.openstreetmap.josm.plugins.pt_assistant.data.DerivedDataSet;
 import org.openstreetmap.josm.plugins.pt_assistant.gui.linear.RelationAccess;
-import org.openstreetmap.josm.plugins.pt_assistant.gui.stoparea.IncompleteMembersWarningPanel;
 import org.openstreetmap.josm.plugins.pt_assistant.gui.stoparea.StopVicinityPanel;
 import org.openstreetmap.josm.plugins.pt_assistant.gui.utils.AbstractVicinityPanel;
+import org.openstreetmap.josm.plugins.pt_assistant.gui.utils.IncompleteMembersWarningPanel;
 import org.openstreetmap.josm.plugins.pt_assistant.gui.utils.UnBoldLabel;
 import org.openstreetmap.josm.plugins.pt_assistant.gui.utils.ZoomSaver;
+import org.openstreetmap.josm.plugins.pt_assistant.utils.DownloadUtils;
 import org.openstreetmap.josm.plugins.pt_assistant.utils.StopUtils;
 
-public class RoutingPanel extends AbstractVicinityPanel {
+public class RoutingPanel extends AbstractVicinityPanel<RoutingDerivedDataSet> {
 
     public static final String TAG_PART_OF_ACTIVE_ROUTE = "partOfActiveRoute";
     public static final String TAG_PART_OF_ACTIVE_ROUTE_VALUE_FORWARD = "forward";
@@ -54,97 +49,7 @@ public class RoutingPanel extends AbstractVicinityPanel {
     public static final String ACTIVE_RELATION_STOP_OFFSET = "activeRelationStopOffset";
 
     public RoutingPanel(IRelationEditorActionAccess editorAccess, ZoomSaver zoom) {
-        super(new DerivedDataSet(editorAccess.getEditor().getLayer().getDataSet()) {
-            @Override
-            protected void addAdditionalGeometry(AdditionalGeometryAccess addTo) {
-                RelationAccess relation = RelationAccess.of(editorAccess);
-
-                relation.getMembers()
-                    .forEach(member -> {
-                        if (member.isNode()) {
-                            addOrGetDerived(withStopOrPlatformFlag(new Node(member.getNode()), member.getRole()));
-                        } else if (member.isWay()) {
-                            addOrGetDerived(withStopOrPlatformFlag(new Way(member.getWay()), member.getRole()));
-                        } else if (member.isRelation()) {
-                            addOrGetDerived(withStopOrPlatformFlag(new Relation(member.getRelation()), member.getRole()));
-                        }
-                    });
-
-                // Now we mark the route geometry as actual driveable way.
-                RouteTraverser routeTraverser = new RouteTraverser(relation);
-                List<RouteSegment> segments = routeTraverser.getSegments();
-                for (int i = 0; i < segments.size(); i++) {
-                    RouteSegment segment = segments.get(i);
-                    // Highlights all segments currently in the route
-                    segment.getWays().forEach(way -> {
-                        OsmPrimitive p = addOrGetDerived(way.getWay());
-                        if (p.hasTag(TAG_PART_OF_ACTIVE_ROUTE)) {
-                            p.put(TAG_PART_OF_ACTIVE_ROUTE, TAG_PART_OF_ACTIVE_ROUTE_VALUE_MULTIPLE);
-                        } else {
-                            if (way.isForward()) {
-                                p.put(TAG_PART_OF_ACTIVE_ROUTE, TAG_PART_OF_ACTIVE_ROUTE_VALUE_FORWARD);
-                            } else {
-                                p.put(TAG_PART_OF_ACTIVE_ROUTE, TAG_PART_OF_ACTIVE_ROUTE_VALUE_BACKWARD);
-                            }
-                        }
-                    });
-                    // Mark the gaps in the line / end of the line
-                    // normal => everything is fine
-                    // broken => there should not be a gap there.
-                    OsmPrimitive first = addOrGetDerived(segment.getWays().get(0).firstNode());
-                    first.put(TAG_ACTIVE_RELATION_SEGMENT_STARTS, i == 0 ? TAG_ACTIVE_RELATION_SEGMENT_VALUE_NORMAL : TAG_ACTIVE_RELATION_SEGMENT_VALUE_BROKEN);
-                    OsmPrimitive last = addOrGetDerived(segment.getWays().get(segment.getWays().size() - 1).lastNode());
-                    last.put(TAG_ACTIVE_RELATION_SEGMENT_ENDS, i == segments.size() - 1 ? TAG_ACTIVE_RELATION_SEGMENT_VALUE_NORMAL : TAG_ACTIVE_RELATION_SEGMENT_VALUE_BROKEN);
-                }
-
-                // Stops on the route
-                List<RouteStop> stops = routeTraverser.findStopPositions();
-                double lastOffsetOnRoute = 0;
-                for (RouteStop stop : stops) {
-                    OsmPrimitive stopNode = addOrGetDerived(stop.getRelationMembers().get(0));
-                    stop.getStopAttributes().getKeys().forEach((k, v) -> stopNode.put("activeRelationStop_" + k, v));
-                    stopNode.put(TAG_ACTIVE_RELATION_STOP_INDEX, stop.getStopIndex() + "");
-                    if (stop instanceof RouteStopPositionOnWay) {
-                        stopNode.put(ACTIVE_RELATION_STOP_OFFSET,
-                            "" + (int) Math.round(((RouteStopPositionOnWay) stop).getPosition().getOffsetInRoute()));
-
-                        PointOnRoute myOffsetOnRoute = ((RouteStopPositionOnWay) stop).getPosition();
-                        if (myOffsetOnRoute.getOffsetInRoute() < lastOffsetOnRoute - 10) { // <10m error
-                            stopNode.put(TAG_ACTIVE_RELATION_STOP_MISORDERED, "1");
-                        }
-                        lastOffsetOnRoute = myOffsetOnRoute.getOffsetInRoute();
-                    } else {
-                        stopNode.put(TAG_ACTIVE_RELATION_STOP_TOO_FAR, "1");
-                        // Do not change lastOffsetOnRoute, so that next stop might get error
-                    }
-                }
-            }
-
-            private OsmPrimitive withStopOrPlatformFlag(OsmPrimitive p, String role) {
-                if (OSMTags.ROUTE_SEGMENT_PT_ROLES.contains(role)) {
-                    p.put(TAG_MEMBER_OF_ACTIVE_RELATION, TAG_MEMBER_OF_ACTIVE_RELATION_VALUE_ROUTE);
-                } else if (OSMTags.PLATFORM_ROLES.contains(role)) {
-                    p.put(TAG_MEMBER_OF_ACTIVE_RELATION, TAG_MEMBER_OF_ACTIVE_RELATION_VALUE_PLATFORM);
-                } else if (OSMTags.STOP_ROLE.contains(role)) {
-                    p.put(TAG_MEMBER_OF_ACTIVE_RELATION, TAG_MEMBER_OF_ACTIVE_RELATION_VALUE_STOP);
-                } else {
-                    p.put(TAG_MEMBER_OF_ACTIVE_RELATION, TAG_MEMBER_OF_ACTIVE_RELATION_VALUE_UNKNOWN);
-                }
-                return p;
-            }
-
-            @Override
-            protected boolean isIncluded(OsmPrimitive primitive) {
-                if (primitive instanceof Way) {
-                    return primitive.hasTag("highway")
-                        || primitive.hasTag("railway");
-                } else if (primitive instanceof Relation) {
-                    return primitive.hasTag("type", "route");
-                } else {
-                    return false;
-                }
-            }
-        }, editorAccess, zoom);
+        super(new RoutingDerivedDataSet(editorAccess), editorAccess, zoom);
 
         // TODO: To actually determine some features, we might need more.
         if (RelationAccess.of(editorAccess)
@@ -168,7 +73,8 @@ public class RoutingPanel extends AbstractVicinityPanel {
         return getOsmPrimitiveAt(point, primitive -> {
             return primitive.hasTag(TAG_ACTIVE_RELATION_SEGMENT_ENDS)
                 || primitive.hasTag(TAG_ACTIVE_RELATION_SEGMENT_STARTS)
-                || primitive.hasTag(TAG_ACTIVE_RELATION_STOP_INDEX);
+                || primitive.hasTag(TAG_ACTIVE_RELATION_STOP_INDEX)
+                || primitive.hasTag(TAG_PART_OF_ACTIVE_ROUTE);
         });
     }
 
@@ -176,21 +82,32 @@ public class RoutingPanel extends AbstractVicinityPanel {
     protected void doAction(Point point, OsmPrimitive derivedPrimitive, OsmPrimitive originalPrimitive) {
         JPopupMenu menu = new JPopupMenu();
 
-        // May have both: Start and end.
-        if (derivedPrimitive.hasTag(TAG_ACTIVE_RELATION_SEGMENT_STARTS)) {
-            if (derivedPrimitive.hasTag(TAG_ACTIVE_RELATION_SEGMENT_STARTS, TAG_ACTIVE_RELATION_SEGMENT_VALUE_BROKEN)) {
-                menu.add(pad(new UnBoldLabel(tr("The route is broken before this point"))));
-            } else {
-                menu.add(pad(new UnBoldLabel(tr("The route starts at this point"))));
+        if (derivedPrimitive.hasTag(TAG_ACTIVE_RELATION_SEGMENT_ENDS) || derivedPrimitive.hasTag(TAG_ACTIVE_RELATION_SEGMENT_STARTS)) {
+            // May have both: Start and end.
+            if (derivedPrimitive.hasTag(TAG_ACTIVE_RELATION_SEGMENT_STARTS)) {
+                if (derivedPrimitive.hasTag(TAG_ACTIVE_RELATION_SEGMENT_STARTS, TAG_ACTIVE_RELATION_SEGMENT_VALUE_BROKEN)) {
+                    menu.add(pad(new UnBoldLabel(tr("The route is broken before this point"))));
+                } else {
+                    menu.add(pad(new UnBoldLabel(tr("The route starts at this point"))));
+                }
             }
-            menu.add(new JSeparator());
-        }
-        if (derivedPrimitive.hasTag(TAG_ACTIVE_RELATION_SEGMENT_ENDS)) {
-            if (derivedPrimitive.hasTag(TAG_ACTIVE_RELATION_SEGMENT_ENDS, TAG_ACTIVE_RELATION_SEGMENT_VALUE_BROKEN)) {
-                menu.add(pad(new UnBoldLabel(tr("The route is broken after this point"))));
-            } else {
-                menu.add(pad(new UnBoldLabel(tr("The route ends at this point"))));
+            if (derivedPrimitive.hasTag(TAG_ACTIVE_RELATION_SEGMENT_ENDS)) {
+                if (derivedPrimitive.hasTag(TAG_ACTIVE_RELATION_SEGMENT_ENDS, TAG_ACTIVE_RELATION_SEGMENT_VALUE_BROKEN)) {
+                    menu.add(pad(new UnBoldLabel(tr("The route is broken after this point"))));
+                } else {
+                    menu.add(pad(new UnBoldLabel(tr("The route ends at this point"))));
+                }
             }
+
+            menu.add(new JMenuItem(new JosmAction(tr("Download adjacent ways"), null, null, null, false) {
+                @Override
+                public void actionPerformed(ActionEvent e) {
+                    DownloadUtils.downloadUsingOverpass("org/openstreetmap/josm/plugins/pt_assistant/gui/routing/downloadWaysStartingAtNode.query.txt", line -> line
+                        .replace("##SELECTOR##", dataSetCopy.getRouteTraverser().getType().getOverpassFilterForPossibleWays())
+                        .replace("##STARTNODEID##", "" + originalPrimitive.getId()));
+                }
+            }));
+
             menu.add(new JSeparator());
         }
 
@@ -209,6 +126,28 @@ public class RoutingPanel extends AbstractVicinityPanel {
             if (derivedPrimitive.hasTag(TAG_ACTIVE_RELATION_STOP_TOO_FAR)) {
                 menu.add(pad(new UnBoldLabel(tr("This stop is too far away from the route. Consider adding a stop_position."))));
             }
+
+            menu.add(new JMenuItem(new JosmAction(tr("Remove this stop from the route"), null, null, null, false) {
+                @Override
+                public void actionPerformed(ActionEvent e) {
+                    editorAccess.getMemberTableModel().removeMembersReferringTo(Arrays.asList(originalPrimitive));
+                }
+            }));
+
+            menu.add(new JSeparator());
+        }
+
+        if (derivedPrimitive.hasTag(TAG_PART_OF_ACTIVE_ROUTE)) {
+            if (derivedPrimitive.hasTag(TAG_PART_OF_ACTIVE_ROUTE, TAG_PART_OF_ACTIVE_ROUTE_VALUE_MULTIPLE)) {
+                menu.add(pad(new UnBoldLabel(tr("This way is included in the route multiple times"))));
+            }
+
+            menu.add(new JMenuItem(new JosmAction(tr("Remove way from route"), null, null, null, false) {
+                @Override
+                public void actionPerformed(ActionEvent e) {
+                    editorAccess.getMemberTableModel().removeMembersReferringTo(Arrays.asList(originalPrimitive));
+                }
+            }));
         }
 
 
@@ -221,4 +160,5 @@ public class RoutingPanel extends AbstractVicinityPanel {
         c.setBorder(new EmptyBorder(5, 5, 5, 5));
         return c;
     }
+
 }
