+++ src/main/java/org/openstreetmap/josm/plugins/pt_assistant/gui/stoparea/StopVicinityPanel.java
@@ -3,10 +3,8 @@ package org.openstreetmap.josm.plugins.pt_assistant.gui.stoparea;
 import static org.openstreetmap.josm.tools.I18n.tr;
 
 import java.awt.BorderLayout;
-import java.awt.Color;
 import java.awt.Point;
 import java.awt.event.ActionEvent;
-import java.text.MessageFormat;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
@@ -23,16 +21,13 @@ import javax.swing.JMenuItem;
 import javax.swing.JPanel;
 import javax.swing.JPopupMenu;
 import javax.swing.JSeparator;
-import javax.swing.border.CompoundBorder;
 import javax.swing.border.EmptyBorder;
-import javax.swing.border.LineBorder;
 import javax.swing.event.PopupMenuEvent;
 import javax.swing.event.PopupMenuListener;
 
 import org.openstreetmap.josm.actions.AutoScaleAction;
 import org.openstreetmap.josm.actions.AutoScaleAction.AutoScaleMode;
 import org.openstreetmap.josm.actions.JosmAction;
-import org.openstreetmap.josm.data.osm.BBox;
 import org.openstreetmap.josm.data.osm.OsmPrimitive;
 import org.openstreetmap.josm.data.osm.OsmPrimitiveType;
 import org.openstreetmap.josm.data.osm.Relation;
@@ -41,13 +36,11 @@ import org.openstreetmap.josm.gui.MainApplication;
 import org.openstreetmap.josm.gui.dialogs.relation.MemberTableModel;
 import org.openstreetmap.josm.gui.dialogs.relation.RelationEditor;
 import org.openstreetmap.josm.gui.dialogs.relation.actions.IRelationEditorActionAccess;
-import org.openstreetmap.josm.gui.layer.OsmDataLayer;
 import org.openstreetmap.josm.gui.mappaint.Cascade;
 import org.openstreetmap.josm.plugins.customizepublictransportstop.OSMTags;
-import org.openstreetmap.josm.plugins.pt_assistant.data.DerivedDataSet;
 import org.openstreetmap.josm.plugins.pt_assistant.gui.linear.RelationAccess;
-import org.openstreetmap.josm.plugins.pt_assistant.gui.linear.RelationEditorAccessUtils;
 import org.openstreetmap.josm.plugins.pt_assistant.gui.utils.AbstractVicinityPanel;
+import org.openstreetmap.josm.plugins.pt_assistant.gui.utils.IncompleteMembersWarningPanel;
 import org.openstreetmap.josm.plugins.pt_assistant.gui.utils.UnBoldLabel;
 import org.openstreetmap.josm.plugins.pt_assistant.gui.utils.ZoomSaver;
 import org.openstreetmap.josm.plugins.pt_assistant.utils.DialogUtils;
@@ -61,14 +54,13 @@ import org.openstreetmap.josm.plugins.pt_assistant.utils.StopUtils;
  * - add / remove platform
  * - add / remove stop_position
  */
-public class StopVicinityPanel extends AbstractVicinityPanel {
+public class StopVicinityPanel extends AbstractVicinityPanel<StopVicinityDerivedDataSet> {
     public static final String CSS_CLASS_PLATFORM = "platform";
     public static final String CSS_CLASS_EMPTYMEMBER = "emptymember";
     public static final String CSS_CLASS_STOP_POSITION = "stop_position";
 
     public StopVicinityPanel(IRelationEditorActionAccess editorAccess, ZoomSaver zoomSaver) {
-        super(createDataSetWithNewRelation(editorAccess.getEditor().getLayer(),
-            editorAccess.getEditor().getRelation(), editorAccess), editorAccess, zoomSaver);
+        super(new StopVicinityDerivedDataSet(editorAccess), editorAccess, zoomSaver);
 
         if (RelationAccess.of(editorAccess)
                 .getMembers()
@@ -88,64 +80,6 @@ public class StopVicinityPanel extends AbstractVicinityPanel {
         );
         legend.setBorder(new EmptyBorder(5, 5, 5, 5));
         add(legend, BorderLayout.SOUTH);
-    }
-
-    private static DerivedDataSet createDataSetWithNewRelation(OsmDataLayer layer, Relation stopRelation,
-                                                               IRelationEditorActionAccess editorAccess) {
-        long editedRelationId = stopRelation == null ? 0 : stopRelation.getId();
-        BBox bBox = new BBox();
-        RelationEditorAccessUtils.streamMembers(editorAccess)
-            // Extra space: Something around 200.500m depending on where we are on the map.
-            .forEach(p -> bBox.addPrimitive(p.getMember(), 0.005));
-
-        return new DerivedDataSet(layer.getDataSet()) {
-            @Override
-            protected boolean isIncluded(OsmPrimitive primitive) {
-                return primitive.getType() != OsmPrimitiveType.RELATION
-                    // Normal primitives: all in bbox
-                    ? bBox.intersects(primitive.getBBox())
-                    // Relations: all except the one we edit
-                    // todo: restrict this, e.g. only PT relations + multipolygons in bbox
-                    : primitive.getId() != editedRelationId;
-            }
-
-            @Override
-            protected void addAdditionalGeometry(AdditionalGeometryAccess addTo) {
-                // Now apply the relation editor changes
-                // Simulate org.openstreetmap.josm.gui.dialogs.relation.actions.SavingAction.applyChanges
-                Relation relation = new Relation();
-                editorAccess.getTagModel().applyToPrimitive(relation);
-                // This is a hack to tag our currently active relation.
-                // There is no id selector in MapCSS, so we need a way to uniquely identify our relation
-                relation.put("activePtRelation", "1");
-
-                if (stopRelation != null) {
-                    addTo.addAsCopy(stopRelation, relation);
-
-                    // Now we search for all sibling relations.
-                    // Due to https://josm.openstreetmap.de/ticket/6129#comment:24 we cannot do it in MapCSS
-                    stopRelation.getReferrers()
-                        .stream()
-                        .filter(r -> r instanceof Relation && StopUtils.isStopAreaGroup((Relation) r))
-                        .flatMap(parent -> ((Relation) parent).getMembers().stream())
-                        .map(RelationMember::getMember)
-                        .filter(sibling -> sibling != stopRelation)
-                        .filter(sibling -> sibling instanceof Relation) // < some group relations contain invalid members
-                        .forEach(sibling -> {
-                            Relation copy = new Relation((Relation) sibling);
-                            copy.put("siblingOfActive", "1");
-                            // This will add the copy with the fake tag.
-                            addOrGetDerived(copy);
-                        });
-
-                } else {
-                    addTo.add(relation);
-                }
-
-                RelationEditorAccessUtils.getRelationMembers(editorAccess)
-                    .forEach(m -> relation.addMember(addOrGetDerivedMember(m)));
-            }
-        };
     }
